+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 - Add plugin support. The manifest may now have a `plugins` section.
 - Add include directories and preprocessor defines support. These can now be specified for source groups.
 - Add `overrides` section to config.
+- Add target specifications for source files.
 
 ### Changed
 - Output of `bender sources` is now JSON formatted.